requirements/dev-requirements-py310.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/dev-requirements-py310.txt requirements/dev-requirements-py310.in
#
alabaster==0.7.12
    # via sphinx
attrs==21.2.0
    # via
    #   hypothesis
    #   jsonschema
    #   pytest
babel==2.9.1
    # via sphinx
backcall==0.2.0
    # via ipython
backports.entry-points-selectable==1.1.1
    # via virtualenv
black[jupyter]==21.10b0
    # via -r requirements/dev-requirements.in
bleach==4.1.0
    # via
    #   nbconvert
    #   readme-renderer
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   black
    #   click-log
    #   python-semantic-release
click-log==0.3.2
    # via python-semantic-release
colorama==0.4.4
    # via twine
commonmark==0.9.1
    # via recommonmark
coverage[toml]==6.1.2
    # via pytest-cov
cryptography==35.0.0
    # via secretstorage
debugpy==1.5.1
    # via ipykernel
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distlib==0.3.3
    # via virtualenv
docutils==0.17.1
    # via
    #   nbsphinx
    #   readme-renderer
    #   recommonmark
    #   sphinx
    #   sphinx-rtd-theme
dotty-dict==1.3.0
    # via python-semantic-release
entrypoints==0.3
    # via
    #   jupyter-client
    #   nbconvert
filelock==3.3.2
    # via
    #   tox
    #   virtualenv
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via python-semantic-release
hypothesis==6.24.3
    # via -r requirements/dev-requirements.in
idna==3.3
    # via requests
imagesize==1.3.0
    # via sphinx
importlib-metadata==4.8.2
    # via
    #   keyring
    #   twine
iniconfig==1.1.1
    # via pytest
invoke==1.6.0
    # via python-semantic-release
ipykernel==6.5.0
    # via -r requirements/dev-requirements.in
ipython==7.29.0
    # via
    #   -r requirements/dev-requirements.in
    #   black
    #   ipykernel
ipython-genutils==0.2.0
    # via nbformat
jedi==0.18.0
    # via ipython
jeepney==0.7.1
    # via
    #   keyring
    #   secretstorage
jinja2==3.0.3
    # via
    #   nbconvert
    #   nbsphinx
    #   sphinx
jsonschema==4.2.1
    # via nbformat
jupyter-client==7.0.6
    # via
    #   ipykernel
    #   nbclient
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.1.2
    # via nbconvert
keyring==23.2.1
    # via twine
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
mypy-extensions==0.4.3
    # via black
nbclient==0.5.8
    # via nbconvert
nbconvert==6.3.0
    # via nbsphinx
nbformat==5.1.3
    # via
    #   nbclient
    #   nbconvert
    #   nbsphinx
nbsphinx==0.8.7
    # via -r requirements/dev-requirements.in
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
oauthlib==3.1.1
    # via requests-oauthlib
packaging==21.2
    # via
    #   bleach
    #   pytest
    #   setuptools-scm
    #   sphinx
    #   tox
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pkginfo==1.7.1
    # via twine
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
prompt-toolkit==3.0.22
    # via ipython
psutil==5.8.0
    # via -r requirements/dev-requirements.in
ptyprocess==0.7.0
    # via pexpect
py==1.11.0
    # via
    #   pytest
    #   tox
pycparser==2.21
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
    #   readme-renderer
    #   sphinx
pyparsing==2.4.7
    # via
    #   -c requirements/requirements-py310.txt
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.5
    # via
    #   -r requirements/dev-requirements.in
    #   pytest-cov
pytest-cov==3.0.0
    # via -r requirements/dev-requirements.in
python-dateutil==2.8.2
    # via
    #   -c requirements/requirements-py310.txt
    #   jupyter-client
python-gitlab==2.10.1
    # via python-semantic-release
python-semantic-release==7.19.2
    # via -r requirements/dev-requirements.in
pytz==2021.3
    # via
    #   -c requirements/requirements-py310.txt
    #   babel
pyzmq==22.3.0
    # via jupyter-client
readme-renderer==30.0
    # via twine
recommonmark==0.7.1
    # via -r requirements/dev-requirements.in
regex==2021.11.10
    # via black
requests==2.26.0
    # via
    #   python-gitlab
    #   python-semantic-release
    #   requests-oauthlib
    #   requests-toolbelt
    #   sphinx
    #   tweepy
    #   twine
requests-oauthlib==1.3.0
    # via tweepy
requests-toolbelt==0.9.1
    # via
    #   python-gitlab
    #   twine
rfc3986==1.5.0
    # via twine
secretstorage==3.3.1
    # via keyring
semver==2.13.0
    # via python-semantic-release
setuptools-scm==6.3.2
    # via dotty-dict
six==1.16.0
    # via
    #   -c requirements/requirements-py310.txt
    #   bleach
    #   python-dateutil
    #   tox
    #   virtualenv
smmap==5.0.0
    # via gitdb
snowballstemmer==2.1.0
    # via sphinx
sortedcontainers==2.4.0
    # via hypothesis
sphinx==4.3.0
    # via
    #   -r requirements/dev-requirements.in
    #   nbsphinx
    #   recommonmark
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements/dev-requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
testpath==0.5.0
    # via nbconvert
tokenize-rt==4.2.1
    # via black
toml==0.10.2
    # via
    #   pytest
    #   tox
tomli==1.2.2
    # via
    #   black
    #   coverage
    #   setuptools-scm
tomlkit==0.7.0
    # via python-semantic-release
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
tox==3.24.4
    # via -r requirements/dev-requirements.in
tqdm==4.62.3
    # via twine
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   nbsphinx
tweepy==4.3.0
    # via -r requirements/dev-requirements.in
twine==3.6.0
    # via python-semantic-release
typing-extensions==3.10.0.2
    # via black
urllib3==1.26.7
    # via requests
virtualenv==20.10.0
    # via tox
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wheel==0.37.0
    # via python-semantic-release
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
